"Small footprint" WP-rule for Lea
The general wp-rule for Lea now only requires the ownership over the
registers that are needed.

This allows reimplementing the existing specific wp-rules on top of
the general wp-rule.

We also delete the wp-rules for failure cases, as they were only used
in the FTLR, which now uses the general rule.